+++ backend/src/main/java/com/hexavolt/backend/repository/StationLocationRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.hexavolt.backend.entity.StationLocation;
 
-public interface StationLocationRepository extends JpaRepository<StationLocation, Integer> {
+public interface StationLocationRepository extends JpaRepository<StationLocation, Long> {
     
 }
