+++ backend/src/main/java/com/hexavolt/backend/repository/WeeklyScheduleRepository.java
@@ -6,8 +6,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.hexavolt.backend.entity.WeeklySchedule;
 
-public interface WeeklyScheduleRepository extends JpaRepository<WeeklySchedule, Integer> {
-  List<WeeklySchedule> findByChargingStationId(Integer stationId);
+public interface WeeklyScheduleRepository extends JpaRepository<WeeklySchedule, Long> {
+  List<WeeklySchedule> findByChargingStationId(Long stationId);
   // List<WeeklySchedule> findByStation_Id(Integer stationId);
 }
